using the error methods we wrote earlier in the handlers and router itself and change the output of ReadNthIDParam from int to int64 to it matches the type of id in movie struct and we don't need the casting.

+++ cmd/api/helpers.go
@@ -8,13 +8,13 @@ import (
 )
 
 // Function will get the specified parameter and then convert it into integer and returns that.
-func (app *application) ReadNthIDParam(r *http.Request, index int) (int, error) {
+func (app *application) ReadNthIDParam(r *http.Request, index int) (int64, error) {
 	param := getParam(r, index)
 	id, err := strconv.Atoi(param)
 	if err != nil {
 		return 0, errors.New("invalid id parameter")
 	}
-	return id, nil
+	return int64(id), nil
 }
 
 type envelope map[string]interface{}
